Minor tweaks to make tests pass:  - Doing @autorelease in test that instantiates lots of rootViewControllers in a loop.  - Adjusting function that creates image from view controller to make memory allocation more disciplined. Not sure if this change actually mattered.
For reasons I still don't understand, a loop that creates a rootViewController and accesses its view will consume all of the system memory in 40-60 iterations. It seems like everything from each iteration should be discarded (especially when encased in an @autorelease directive), and indeed this seems to happen during normal operation, but the loop still brings down the system.